requirements.txt
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
defusedxml==0.7.1
Deprecated==1.2.18
Django==5.2
django-cors-headers==4.7.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-users==14.0.1
greenlet==3.1.1
h11==0.14.0
httpie==3.2.4
idna==3.10
kombu==5.5.2
limits==4.6
makefun==1.15.6
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.3.2
mysqlclient==2.2.7
packaging==24.2
passlib==1.7.4
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pwdlib==0.2.1
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
setuptools==78.1.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
